# guangdong-energy-monitor/docs/API_CONTRACT.yaml
openapi: 3.0.3
info:
  title: 广东省电厂监控平台 API 契约
  description: |
    华电集团广东电厂监控平台后端接口契约文档。
    所有接口统一通过 Gateway（8080）访问，需携带 JWT Token。
    所有响应统一使用 R<T> 包装结构。
    包含认证授权、用户管理、角色管理、组织架构、电厂管理、水文站管理、
    GIS 空间服务、报警管理、预警管理、规则管理、统计分析、预测服务、数据采集共13个控制器、54个REST端点。
  version: 1.0.0
  contact:
    name: 华电集团广东电厂监控平台研发组
servers:
  - url: http://localhost:8080
    description: 本地开发环境（Gateway 入口）
  - url: https://api.gd-energy.example.com
    description: 生产环境
security:
  - BearerAuth: []
...
paths:
...
  /api/users:
    get:
      tags: [用户管理]
      summary: 用户列表（分页查询）
      description: 分页查询用户列表，支持关键字搜索
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: keyword
          in: query
          description: 用户名/姓名关键字
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResultSysUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [用户管理]
      summary: 新建用户
      description: 创建新用户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysUser'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SysUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [用户管理]
      summary: 更新用户
      description: 更新用户信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysUser'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SysUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{id}:
    get:
      tags: [用户管理]
      summary: 用户详情
      description: 根据用户ID查询详情
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [用户管理]
      summary: 删除用户
      description: 根据用户ID删除用户
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/R'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # 角色管理 (RoleController) - 5 endpoints
  # ============================================================
  /api/roles:
    get:
      tags: [角色管理]
      summary: 角色列表
      description: 查询所有角色列表
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SysRole'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [角色管理]
      summary: 新建角色
      description: 创建新角色
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysRole'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SysRole'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [角色管理]
      summary: 更新角色
      description: 更新角色信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysRole'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SysRole'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/roles/{id}:
    get:
      tags: [角色管理]
      summary: 角色详情
      description: 根据角色ID查询详情
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SysRole'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [角色管理]
      summary: 删除角色
      description: 根据角色ID删除角色
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/R'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # 组织架构 (OrgController) - 6 endpoints
  # ============================================================
  /api/orgs:
    get:
      tags: [组织架构]
      summary: 组织列表
      description: 查询所有组织列表（平铺）
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SysOrg'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [组织架构]
      summary: 新建组织
      description: 创建新组织节点
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysOrg'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SysOrg'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [组织架构]
      summary: 更新组织
      description: 更新组织节点信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysOrg'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SysOrg'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/orgs/tree:
    get:
      tags: [组织架构]
      summary: 组织树形结构
      description: 查询组织树形结构（带 children）
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SysOrg'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/orgs/{id}:
    get:
      tags: [组织架构]
      summary: 组织详情
      description: 根据组织ID查询详情
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SysOrg'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [组织架构]
      summary: 删除组织
      description: 根据组织ID删除组织节点
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/R'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # 电厂管理 (PlantController) - 6 endpoints
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer Token 认证，登录后获取 token，请求头携带 Authorization: Bearer <token>"
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码（从1开始）
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: 每页条数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdParam:
      name: id
      in: path
      required: true
      description: 主键ID
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: 请求参数错误（400 PARAM_ERROR）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
    Unauthorized:
      description: 未认证或Token无效（401 UNAUTHORIZED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
...
    Forbidden:
      description: 无权限访问（403 FORBIDDEN）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
    NotFound:
      description: 资源不存在（404 NOT_FOUND）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
...
  schemas:
...
    R:
      type: object
      description: |
        统一响应包装结构 R<T>。
        错误码：200 SUCCESS, 400 PARAM_ERROR, 401 UNAUTHORIZED,
        402 TOKEN_EXPIRED, 403 FORBIDDEN, 404 NOT_FOUND,
        500 BUSINESS_ERROR, 5000 SYSTEM_ERROR
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 200
        msg:
          type: string
          description: 提示信息
          example: 操作成功
        data:
          description: 业务数据（具体结构见各接口）
          nullable: true

    # --------------------------------------------------------
    # 分页结果 PageResult<T>
    # --------------------------------------------------------
    PageResult:
      type: object
      description: 分页结果通用结构 PageResult<T>
      properties:
        list:
          type: array
          description: 数据列表
          items: {}
        total:
          type: integer
          format: int64
          description: 总记录数
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页条数
    PageResultSysUser:
      allOf:
        - $ref: '#/components/schemas/PageResult'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/SysUser'
...
    LoginVO:
      type: object
      description: 登录返回视图对象
      properties:
        token:
          type: string
          description: JWT Token
        userId:
          type: integer
          format: int64
          description: 用户ID
        username:
          type: string
          description: 用户名
        realName:
          type: string
          description: 真实姓名
        roles:
          type: array
          description: 角色列表
          items:
            $ref: '#/components/schemas/UserRole'
        permissions:
          type: array
          description: 权限标识列表
          items:
            type: string
    SysUser:
      type: object
      description: 系统用户
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        username:
          type: string
          description: 用户名
        realName:
          type: string
          description: 真实姓名
        phone:
          type: string
          description: 手机号
        email:
          type: string
          description: 邮箱
        orgId:
          type: integer
          format: int64
          description: 所属组织ID
        status:
          type: integer
          description: 状态（0禁用 1启用）
        createTime:
          type: string
          format: date-time
          description: 创建时间
        updateTime:
          type: string
          format: date-time
          description: 更新时间

    # --------------------------------------------------------
    # 角色
    # --------------------------------------------------------
    SysRole:
      type: object
      description: 系统角色
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        name:
          type: string
          description: 角色名称
        code:
          type: string
          description: 角色编码
        description:
          type: string
          description: 角色描述
        createTime:
          type: string
          format: date-time
          description: 创建时间

    # --------------------------------------------------------
    # 组织架构
    # --------------------------------------------------------
    SysOrg:
      type: object
      description: 组织架构（支持树形结构，带 children）
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        parentId:
          type: integer
          format: int64
          description: 父级ID
        name:
          type: string
          description: 组织名称
        code:
          type: string
          description: 组织编码
        sort:
          type: integer
          description: 排序号
        createTime:
          type: string
          format: date-time
          description: 创建时间
        children:
          type: array
          description: 子组织列表
          items:
            $ref: '#/components/schemas/SysOrg'

    # --------------------------------------------------------
    # 电厂
    # --------------------------------------------------------
...
    UserRole:
      type: string
      description: 用户角色
      enum:
        - ADMIN
        - PROD_SAFETY
        - PLANT_MANAGER
        - OPERATOR
...